refactor: create new folders
