+++ src/pieces/knight.rs
@@ -1,6 +1,6 @@
 use super::{Movable, PieceColor, PieceMove, Position};
-use crate::board::{Coord, GameBoard};
 use crate::constants::DisplayMode;
+use crate::game::board::{Coord, GameBoard};
 use crate::utils::{cleaned_positions, impossible_positions_king_checked, is_cell_color_ally};
 pub struct Knight;
