+++ src/pieces/rook.rs
@@ -1,5 +1,5 @@
 use super::{Movable, PieceColor, PieceMove, Position};
-use crate::board::{Coord, GameBoard};
+use crate::game::board::{Coord, GameBoard};
 use crate::constants::DisplayMode;
 use crate::utils::{
     cleaned_positions, get_piece_color, impossible_positions_king_checked, is_cell_color_ally,
